Draggable: Move draggable into widgets folder
